% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Check Inputs for Linear Programming Problem}
\usage{
check_inputs(c, A, b, slack)
}
\arguments{
\item{c}{A numeric vector of coefficients for the objective function.}

\item{A}{A numeric matrix of coefficients for the constraints.}

\item{b}{A numeric vector of constants on the right-hand side of the constraints.}

\item{slack}{A numeric vector of indices for slack variables (forming the identity matrix).}
}
\description{
This function checks the validity of inputs for a linear programming problem.
}
