% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPsimplex_parallel.R
\name{LPsimplex_parallel}
\alias{LPsimplex_parallel}
\title{LPsimplex_parallel}
\usage{
LPsimplex_parallel(c, A, b, slack, max = TRUE, process = TRUE)
}
\arguments{
\item{c}{A numeric vector of coefficients for the objective function.}

\item{A}{A numeric matrix of coefficients for the constraints (after adding slack variables).}

\item{b}{A numeric vector of constants on the right-hand side of the constraints.}

\item{slack}{A numeric vector of indices for slack variables (forming the identity matrix).}

\item{max}{A logical value indicating whether to maximize the objective function.
- `TRUE`: Maximize the objective function.
- `FALSE`: Minimize the objective function.}

\item{process}{A logical value indicating whether to print the simplex table during the solving process.
- `TRUE`: Print the simplex table at each step.
- `FALSE`: Do not print the simplex table in the process.}
}
\value{
A list containing the optimal solution and the optimal value of the objective function.
}
\description{
This function solves a linear programming problem using the simplex method with parallelization.
}
\examples{
c=c(60,120,0,0,0)
A=matrix(c(9,3,4,4,10,5,1,0,0,0,1,0,0,0,1),3)
b=c(360,300,200)
slack=c(3,4,5)
result <- LPsimplex_parallel(c, A, b, slack, max = TRUE, process = TRUE)
print(result)
}
