% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_simplex_step.R
\name{parallel_simplex_step}
\alias{parallel_simplex_step}
\title{Parallel Simplex Step}
\usage{
parallel_simplex_step(c, A, b, B, pcol, prow, process)
}
\arguments{
\item{c}{A numeric vector of coefficients for the objective function.}

\item{A}{A numeric matrix of coefficients for the constraints.}

\item{b}{A numeric vector of constants on the right-hand side of the constraints.}

\item{B}{A numeric vector of indices for basic variables.}

\item{pcol}{The index of the entering variable (pivot column).}

\item{prow}{The index of the pivot row.}

\item{process}{A logical value indicating whether to print the simplex table.}
}
\value{
A list containing updated values of `A`, `b`, `sigma`, and `B`.
}
\description{
This function performs key steps of the simplex method in parallel, with cross-platform support.
}
