% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_simplex_table.R
\name{print_simplex_table}
\alias{print_simplex_table}
\title{Print Simplex Table}
\usage{
print_simplex_table(CB, Base, b, A, sigma, title)
}
\arguments{
\item{CB}{A numeric vector representing the coefficients of the basic variables in the objective function}

\item{Base}{A numeric vector indicating the indices of the basic variables in the current simplex tableau}

\item{b}{A numeric vector representing the right-hand side values of the constraints (the constants in the equations)}

\item{A}{A numeric matrix representing the coefficients of the variables in the constraint (including slack variables)}

\item{sigma}{A numeric vector representing the reduced costs of the non-basic variables in the current simplex tableau}

\item{title}{A character string providing a title or description for the current simplex tableau}
}
\description{
This function prints the simplex table of a certain step
}
